requirements.txt
# pip install -r "requirements.txt"

influxalchemy
influxdb
pandas
python-docx
pytz
requests
